requirements.txt
pandas
streamlit
plotly
openpyxl
numpy
bcrypt
supabase
python-dotenv
